src/main/java/dat3/rename_me/api/CategoryController.java
package dat3.rename_me.api;

import dat3.rename_me.dto.CategoryRequest;
import dat3.rename_me.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/categories")
public class CategoryController {
    private final CategoryService categoryService;


    public CategoryController(CategoryService categoryService) {
        this.categoryService = categoryService;
    }

    @GetMapping
    public List<CategoryDTO> getAllCategories() {
        return categoryService.getAllCategories();
    }

    @PostMapping
    public void addCategory(@RequestBody CategoryRequest categoryRequest) {
        CategoryDTO categoryDTO = new CategoryDTO(0, categoryRequest.getName());
        categoryService.addCategory(categoryDTO);
    }
}
